App/requirements.txt
streamlit
pandas
azure-storage-blob
azure-data-tables
python-dotenv
google-generativeai
plotly
